+++ libs/components/grids/projects/grids/documentation/code-examples/grid/basic/grid-demo.module.ts
@@ -1,0 +1,29 @@
+import {
+  NgModule
+} from '@angular/core';
+
+import {
+  CommonModule
+} from '@angular/common';
+
+import {
+  SkyGridModule
+} from 'projects/grids/src/public-api';
+
+import {
+  GridDemoComponent
+} from './grid-demo.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    SkyGridModule
+  ],
+  declarations: [
+    GridDemoComponent
+  ],
+  exports: [
+    GridDemoComponent
+  ]
+})
+export class GridDemoModule { }