+++ libs/components/grids/src/app/public/plugin-resources/grids-resources-provider.ts
@@ -1,8 +1,0 @@
-import {
-  SkyAppLocaleInfo,
-  SkyLibResourcesProvider
-} from '@skyux/i18n';
-
-export class SkyGridsResourcesProvider implements SkyLibResourcesProvider {
-  public getString: (localeInfo: SkyAppLocaleInfo, name: string) => string;
-}